+++ frontend/app/components/Navigation/VerbaButton.tsx
@@ -21,7 +21,6 @@ interface VerbaButtonProps {
   loading?: boolean;
   text_size?: string;
   icon_size?: number;
-  button_size?: string;
   onClickParams?: any[]; // New prop to pass additional parameters
 }
 
@@ -38,12 +37,11 @@ const VerbaButton: React.FC<VerbaButtonProps> = ({
   selected = false,
   selected_color = "bg-button-verba",
   selected_text_color = "text-text-verba-button",
-  text_size = "text-sm",
-  icon_size = 15,
+  text_size = "text-xs",
+  icon_size = 12,
   type = "button",
   loading = false,
   circle = false,
-  button_size = "",
   onClickParams = [],
 }) => {
   return (
@@ -52,7 +50,7 @@ const VerbaButton: React.FC<VerbaButtonProps> = ({
       key={key}
       className={
         className +
-        ` btn rounded-lg flex-grow items-center justify-center border-none ${circle ? "btn-circle" : ""} ${button_size} hover:bg-button-hover-verba hover:text-text-verba-button ${selected ? selected_color + " shadow-md " + selected_text_color : " bg-button-verba shadow-none text-text-alt-verba-button"} `
+        ` p-3 transition-all active:scale-95 scale-100 duration-300 flex gap-1 items-center justify-center ${circle ? "rounded-full" : "rounded-lg"} hover:bg-button-hover-verba hover:text-text-verba-button ${selected ? selected_color + " shadow-md " + selected_text_color : " bg-button-verba text-text-alt-verba-button"} `
       }
       onClick={(e) => onClick(e, ...onClickParams)}
       disabled={disabled}
@@ -60,17 +58,15 @@ const VerbaButton: React.FC<VerbaButtonProps> = ({
       onMouseLeave={onMouseLeave}
     >
       {loading ? (
-        <span className="text-text-verba-button loading loading-spinner loading-sm"></span>
+        <span className="text-text-verba-button loading loading-spinner loading-xs"></span>
       ) : (
         <>
-          <div className="flex gap-2 items-center">
-            {Icon && <Icon size={icon_size} className="w-[20px]" />}
-            {title && (
-              <p title={title} className={text_size + " " + text_class_name}>
-                {title}
-              </p>
-            )}
-          </div>
+          {Icon && <Icon size={icon_size} className="w-[20px]" />}
+          {title && (
+            <p title={title} className={text_size + " " + text_class_name}>
+              {title}
+            </p>
+          )}
         </>
       )}
     </button>