+++ frontend/tailwind.config.ts
@@ -49,6 +49,16 @@ const config: Config = {
       yellow: colors.yellow,
       red: colors.red,
     },
+    fontSize: {
+      xs: "0.75rem",
+      sm: "0.8rem",
+      base: "1rem",
+      xl: "1.25rem",
+      "2xl": "1.563rem",
+      "3xl": "1.953rem",
+      "4xl": "2.441rem",
+      "5xl": "3.052rem",
+    },
     extend: {
       colors: {
         "bg-verba": "var(--bg-verba, #FEF7F7)",