+++ goldenverba/components/interfaces.py
@@ -391,7 +391,7 @@ class Generator(VerbaComponent):
         super().__init__()
         self.context_window = 5000
         self.config["System Message"] = InputConfig(
-            type="text",
+            type="textarea",
             value="You are Verba, a chatbot for Retrieval Augmented Generation (RAG). You will receive a user query and context pieces that have a semantic similarity to that query. Please answer these user queries only with the provided context. Mention documents you used from the context if you use them to reduce hallucination. If the provided documentation does not provide enough information, say so. If the user asks questions about you as a chatbot specifially, answer them naturally. If the answer requires code examples encapsulate them with ```programming-language-name ```. Don't do pseudo-code.",
             description="System Message",
             values=[],