+++ goldenverba/components/managers.py
@@ -36,7 +36,9 @@ from goldenverba.components.reader.UnstructuredAPI import UnstructuredReader
 from goldenverba.components.reader.AssemblyAIAPI import AssemblyAIReader
 from goldenverba.components.reader.HTMLReader import HTMLReader
 from goldenverba.components.reader.FirecrawlReader import FirecrawlReader
-from goldenverba.components.reader.UpstageDocumentParse import UpstageDocumentParseReader
+from goldenverba.components.reader.UpstageDocumentParse import (
+    UpstageDocumentParseReader,
+)
 
 # Import Chunkers
 from goldenverba.components.chunking.TokenChunker import TokenChunker
@@ -69,6 +71,7 @@ from goldenverba.components.generation.OllamaGenerator import OllamaGenerator
 from goldenverba.components.generation.OpenAIGenerator import OpenAIGenerator
 from goldenverba.components.generation.GroqGenerator import GroqGenerator
 from goldenverba.components.generation.UpstageGenerator import UpstageGenerator
+
 try:
     import tiktoken
 except Exception:
@@ -304,24 +307,31 @@ class WeaviateManager:
             msg.info(
                 f"Collection: {collection_name} does not exist, creating new collection."
             )
-            await client.collections.create(name=collection_name)
-        return True
+            returned_collection = await client.collections.create(name=collection_name)
+            if returned_collection:
+                return True
+            else:
+                return False
+        else:
+            return True
 
     async def verify_embedding_collection(self, client: WeaviateAsyncClient, embedder):
         if embedder not in self.embedding_table:
             self.embedding_table[embedder] = "VERBA_Embedding_" + re.sub(
                 r"[^a-zA-Z0-9]", "_", embedder
             )
-            await self.verify_collection(client, self.embedding_table[embedder])
-        return True
+            return await self.verify_collection(client, self.embedding_table[embedder])
+        else:
+            return True
 
     async def verify_cache_collection(self, client: WeaviateAsyncClient, embedder):
         if embedder not in self.embedding_table:
             self.embedding_table[embedder] = "VERBA_Cache_" + re.sub(
                 r"[^a-zA-Z0-9]", "_", embedder
             )
-            await self.verify_collection(client, self.embedding_table[embedder])
-        return True
+            return await self.verify_collection(client, self.embedding_table[embedder])
+        else:
+            return True
 
     async def verify_embedding_collections(
         self, client: WeaviateAsyncClient, environment_variables, libraries
@@ -804,14 +814,18 @@ class WeaviateManager:
     ):
         if await self.verify_embedding_collection(client, embedder):
             embedder_collection = client.collections.get(self.embedding_table[embedder])
-            weaviate_chunks = await embedder_collection.query.fetch_objects(
-                filters=(
-                    Filter.by_property("doc_uuid").equal(doc_uuid)
-                    & Filter.by_property("chunk_id").contains_any(ids)
-                ),
-                sort=Sort.by_property("chunk_id", ascending=True),
-            )
-            return weaviate_chunks.objects
+            try:
+                weaviate_chunks = await embedder_collection.query.fetch_objects(
+                    filters=(
+                        Filter.by_property("doc_uuid").equal(str(doc_uuid))
+                        & Filter.by_property("chunk_id").contains_any(list(ids))
+                    ),
+                    sort=Sort.by_property("chunk_id", ascending=True),
+                )
+                return weaviate_chunks.objects
+            except Exception as e:
+                msg.fail(f"Failed to fetch chunks: {str(e)}")
+                raise e
 
     ### Suggestion Logic
 
@@ -908,13 +922,16 @@ class WeaviateManager:
                 filters = Filter.by_property("doc_uuid").contains_any(document_uuids)
             else:
                 filters = None
-
-            response = await embedder_collection.aggregate.over_all(
-                filters=filters,
-                group_by=GroupByAggregate(prop="doc_uuid"),
-                total_count=True,
-            )
-            return len(response.groups)
+            try:
+                response = await embedder_collection.aggregate.over_all(
+                    filters=filters,
+                    group_by=GroupByAggregate(prop="doc_uuid"),
+                    total_count=True,
+                )
+                return len(response.groups)
+            except Exception as e:
+                msg.fail(f"Failed to retrieve data count: {str(e)}")
+                return 0
 
     async def get_chunk_count(
         self, client: WeaviateAsyncClient, embedder: str, doc_uuid: str