+++ CHANGELOG.md
@@ -12,8 +12,9 @@ All notable changes to this project will be documented in this file.
 - Added Groq
 - Language Detection for languages like chinese, english, french, german and dutch
 - Improve Markdown Chunking (https://github.com/weaviate/Verba/pull/323)
+- Upgrade to latest Weaviate Client
 
-# Fixed
+## Fixed
 
 - Catch Exception when trying to access the OpenAI API Embedding endpoint to retrieve model names
 - Fixed reading empty string as environment variables
@@ -21,6 +22,8 @@ All notable changes to this project will be documented in this file.
 - Changed Collection names to prevent conflicts with other older Verba version generated collections
 - Ensure Ollama URL is parsed correctly (https://github.com/weaviate/Verba/pull/327)
 - Fixing typos (https://github.com/weaviate/Verba/pull/329)
+- System Message is now a textarea in the frontend
+- Race Condition when multiple requests try to create same client (https://github.com/weaviate/Verba/issues/335)
 
 ## [2.0.0] Importastic